Define KeyValueContextTriple Struct
Add "KeyValueContextTriple"-struct to class-library. By proxy, a class
library - along with a namespace - was created.

KVC-struct will be used to group related variables needed to transfer
data from a sender to reciever.

A KVC can also be seen as a nested-KV,(Key1, (Key2, Value)) where 'Key1'
is needed because Key2 lacks an unique attribute to be used as a proper-
key. 'Key2' will typically be value-data, so this is often the case.