Refactor KeyValueYield to Provide More Applicable Data
Add undefined-indicator "NaN" to Value. Add "TKey" to parameter-list of Process.

Provide more applicable information to accessors, to increase the accuracy and efficiency of potential responses.

The "NaN"-indicator was included to differeniate intentional null-values, from undefined/unknown values.

The "TKey" was included to indicate to the process which proxy it was recieving the data from, which can be used to more-easily fine-tune responses.