Define Glossary Generic-Class
Add "Glossary<TKey, TValue>"-class to class-library. Include public-API definitions.

Glossary is a collection focused on handling resolution-conflicts.

An example of a "resolution-conflict" is a KeyValuePair. The key is resolved by attaining the value.

Glossary takes this resolution one-step further, and includes a process that uses the value. This means that the key is resolved by the value, and the way to use that value is resolved by the process.

This collection is intended to be used in Singleton-Models, to resolve the passing of data between modules. However it can also be used as:

+ A built-in one-way pipe

+ A structured way to process multiple data-inputs from random-sources

+ A way to share data without exposing ANY implementation