Adjust Project License
Change "GPL-3.0.-or-later" license to "LGPL-3.0.-or-later" license.

rThis is make the library viable in more projects. This allows for compliation, editing, parsing, etc software to be made using this library without forcing its constituents to also use GPL.

LGPL can still be "upgraded" to GPL in projects that deem it necessary.